feat(grunt): move build system to webpack
This moves the client files to ./client instead of inside ./public/js, and
also moves our system to webpack instead of relying on everything being in
window. This is good for us because webpack allows us to perform some
optimizations that concat + uglify does not. It also allows for the code
to be better maintained via using 'require' and making the dependencies
explicit rather than implicit.